requirements.txt
groq
python-dotenv
python-dateutil>=2.8.2
aiohttp>=3.9.0
asyncpg>=0.29.0
psycopg[binary]>=3.2.0
pymongo>=4.6.0
duckdb>=0.10.0
anyio<5,>=3.5.0
distro<2,>=1.7.0
httpx<1,>=0.23.0
pydantic<3,>=1.9.0
sniffio
typing-extensions<5,>=4.10
idna>=2.8
certifi
httpcore==1.*
h11>=0.16
annotated-types>=0.6.0
pydantic-core==2.41.5
typing-inspection>=0.4.2
streamlit>=1.28.0
sqlglot>=24.0.0,<28.0.0
gdown>=5.0.0
